pyproject.toml
[tool.poetry]
name = "statuspage-backend"
version = "0.1.0"
description = "A sample of status page application"
authors = ["Chintapallidharmendra <chintapallidharmendra@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.9"
Django = "^5.1.2"
djangorestframework = "^3.15.2"
django-rest-knox = "^5.0.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
